local_server/requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fonttools==4.38.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.23.5
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
packaging==22.0
pendulum==2.1.2
Pillow==9.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytzdata==2020.1
requests==2.28.1
six==1.16.0
urllib3==1.26.13
websockets==10.4
